Change parse_message_line to parse_message
Should parse_message_line ever return None
git2json will fail with a TypeError when
we go to join a list with a None in it.

Since we may want to be able to parse logs
with messageless commit, we still want to
be able to output None after parsing
a commit message sometimes.

Since all parse_message_line really does
is remove whitespace formatting from
the beginning of the commit log line,
we can simply remove this whitespace formatting
from the beginning of all lines. If the result
of this operation is empty, we can record
that commit as having no message.

The alternative was to keep the parse_blah_line
symmetry of parse_message_line and then test
whether the result of joining the list was empty,
having parse_message_line return an empty string.